requirements.txt
streamlit>=1.30
pandas>=2.0
google-generativeai>=0.3.0
openpyxl
tabulate